Initial commit - load all branch devices by parsing BGP table
This script parses the BGP table of a main WAN router. It derrives all of the branch devices (router, firewall, and switch) from each of these BGP prefixes. Then it begins crawling each device and adding it to django with the appropriate attributes. The device attributes are discovered via SSH.